Create Lists collection using Floating Button and store Lists in firestore.
Created a Floating button that takes list name as user input
and creates a new collection of lists.
Also in this collection each document contains data for each
corresponding lists.
The document name is same as the name of the list.
The document contains one field value i.e:- list name.